--- a/docs/stories/epic-1/story-1.7-performance-monitoring.md
+++ b/docs/stories/epic-1/story-1.7-performance-monitoring.md
@@ -1,7 +1,7 @@
 # Story 1.7: Performance Monitoring Framework
 
 ## Status
-Ready for Done
+Done
 
 ## Story