… .github/workflows/performance.yml
… docs/qa/assessments/1.7-nfr-20250909.md
… docs/qa/assessments/1.7-risk-20250909.md
… docs/qa/assessments/1.7-test-design-20250909.md
… docs/qa/assessments/1.7-trace-20250909.md
… docs/qa/gates/1.7-performance-monitoring.yml
… flattened-codebase.stats.md
… package.json
… packages/core/src/monitoring/PerformanceProfiler.ts
… packages/core/src/monitoring/RegressionDetector.ts
… packages/core/src/monitoring/decorators.ts
… packages/core/tests/benchmarks/core.bench.ts
… packages/core/tests/benchmarks/runner.ts
… packages/core/tests/benchmarks/simple.bench.ts
… packages/core/tests/benchmarks/state.bench.ts
… packages/core/tests/benchmarks/workflow.bench.ts
… packages/core/tests/monitoring/PerformanceMonitor.test.ts
… packages/core/tests/monitoring/bottleneck-detection.test.ts
… packages/core/tests/monitoring/dashboard.test.ts
… packages/core/tests/monitoring/decorators.test.ts
… packages/core/tests/monitoring/memory-profiling.test.ts