## Summary

✅ **Story 1.7: Performance Monitoring Framework - COMPLETE**

This PR implements a comprehensive performance monitoring framework for the BMAD Checklist Manager, ensuring all operations meet the <100ms requirement throughout development.

### 🎯 Key Features Delivered

- **Real-time Performance Monitoring**: Complete PerformanceMonitor infrastructure with metrics collection, budget enforcement, and violation detection
- **Memory Profiling**: Comprehensive memory usage tracking with baseline (30MB) and peak (50MB) validation
- **Dashboard & Visualization**: Development-time performance dashboard with real-time metrics display
- **Bottleneck Detection**: Advanced performance analysis with CPU, memory, and duration bottleneck identification
- **CI/CD Integration**: Automated performance testing with regression detection and PR blocking capabilities

### 📊 Quality Metrics Achieved

- **Quality Gate**: ✅ **PASS** (100/100 quality score)
- **Test Coverage**: 666 tests passing (85.88% coverage)
- **Requirements Coverage**: 26/26 acceptance criteria (100%)
- **NFR Compliance**: All PASS (Security, Performance, Reliability, Maintainability)
- **Performance Targets**: All benchmarks validate <100ms operations, <50MB memory usage
- **Technical Debt**: 0 (Zero critical issues)

### 🔧 Implementation Highlights

#### Core Performance Infrastructure
- `PerformanceMonitor.ts` - Central monitoring service with metrics collection
- `PerformanceProfiler.ts` - Memory profiling and performance analysis
- `PerformanceDashboard.ts` - Real-time development dashboard
- `RegressionDetector.ts` - CI/CD performance regression detection
- `decorators.ts` - Performance measurement decorators for easy integration

#### Comprehensive Testing
- **Unit Tests**: 87 tests across all monitoring components
- **Memory Profiling Tests**: 20+ tests validating memory requirements
- **Dashboard Tests**: 15 tests for visualization functionality  
- **Bottleneck Detection Tests**: 22 tests for performance analysis
- **Benchmark Suite**: Complete validation of all performance targets

#### Quality Assurance Documentation
- **Requirements Traceability**: Complete mapping of 26 requirements to tests with Given-When-Then patterns
- **NFR Assessment**: All four core NFRs validated with comprehensive evidence
- **Quality Gate**: PASS decision with detailed audit trail and historical tracking

### 🚀 Performance Targets Validated

All performance requirements comprehensively tested and validated:

- ✅ Command execution <100ms (P95 validation)
- ✅ Application startup <500ms (P95 validation)
- ✅ Template parsing <100ms (P95 validation)  
- ✅ State operations: Save <50ms, Load <30ms
- ✅ TUI frame render <16.67ms (60fps requirement)
- ✅ File system operations <50ms
- ✅ Memory usage <50MB with realistic load simulation

### 🔒 Security & Reliability

- **Secrets Detection**: Comprehensive pattern-based secret scanning with 15+ detection types
- **Error Handling**: 97+ try-catch blocks across 21 source files
- **Resource Management**: Proper cleanup, timer management, WeakMap usage
- **Graceful Degradation**: Performance monitoring can be disabled without breaking functionality

## Test Plan

- [x] All 666 tests passing with 85.88% coverage
- [x] Lint and type checking clean (no warnings/errors)
- [x] Performance benchmarks validate all targets
- [x] Memory profiling tests confirm <50MB usage
- [x] Dashboard functionality verified in development mode
- [x] CI/CD pipeline integration tested with regression detection
- [x] Quality gate validation complete with PASS status

🤖 Generated with [Claude Code](https://claude.ai/code)